+++ com.unity.shadergraph/CHANGELOG.md
@@ -43,6 +43,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Shader Graph now has __Group Node__, where you can group together several nodes. You can use this to keep your Graphs organized and nice.
 
+### Fixed
+- The expanded state of blackboard properties are now remembered during a Unity session.
+
 ## [5.1.0] - 2018-11-19
 ### Added
 - You can now show and hide the Main Preview and the Blackboard from the toolbar.